Closes #8966 - Rename 'integer' parameter type to 'numeric'

+++ app/assets/javascripts/discerner/search_parameter_value.js
@@ -42,8 +42,8 @@ Discerner.SearchParameterValue.UI = function (config) {
         }
         $(config.container).find('.parameter_value, .remove').show();
         $(config.container).find('a.add_search_parameter_values').show();
-      } else {                                                                        // date, text and integer parameters
-        var parameter_classes = ['date', 'integer', 'text']
+      } else {                                                                        // date, text and numeric parameters
+        var parameter_classes = ['date', 'numeric', 'text']
         for (var i in parameter_classes) {
           if ($(selectedParameter).hasClass(parameter_classes[i])) {
             row.find('.operator option:not(.' + parameter_classes[i] +')').detach();
